% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_values.R
\name{check_values}
\alias{check_values}
\title{Check if Column Values are Restricted to Two Specific Values}
\usage{
check_values(data, column_name, value1, value2)
}
\arguments{
\item{data}{A data frame containing the column to check.}

\item{column_name}{A character string specifying the name of the column to check.}

\item{value1}{The first allowed value in the column.}

\item{value2}{The second allowed value in the column.}
}
\value{
A character string. If all values in the column are either \code{value1} or \code{value2}, the function
returns a message confirming this. If there are invalid values, it returns a message listing
the row numbers where these values appear.
}
\description{
This function checks whether all values in a specified column of a data frame are restricted to
two given values. If any values are not one of the allowed values, it returns the row indices where
the invalid values appear. Otherwise, it returns a message confirming that all values are valid.
}
\examples{
check_values(ourdata, "gender", "M", "F")

}
