% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_non_numeric.R
\name{find_non_numeric}
\alias{find_non_numeric}
\title{Identify and Remove Rows with Non-Numeric Values in a Specified Column}
\usage{
find_non_numeric(df, col1)
}
\arguments{
\item{df}{A data frame that contains the column to be checked.}

\item{col1}{A character string specifying the name of the column to inspect.}
}
\value{
A data frame with rows containing non-numeric values removed. If all values
in the specified column are numeric, the function returns the original data frame.
}
\description{
This function scans a given column in a data frame to identify and remove rows where
the values are non-numeric. It attempts to coerce the column values to numeric. If a
value cannot be coerced (resulting in \code{NA}), those rows are excluded from the output.
}
\examples{
find_non_numeric(ourdata, "age")

}
