% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plot.R
\name{generate_plot}
\alias{generate_plot}
\title{Generate a Scatter Plot for Two Variables}
\usage{
generate_plot(data, x_var, y_var)
}
\arguments{
\item{data}{A data frame containing the variables to plot.}

\item{x_var}{A character string specifying the name of the variable to be plotted on the x-axis.}

\item{y_var}{A character string specifying the name of the variable to be plotted on the y-axis.}
}
\value{
This function does not return a value. It creates and displays a scatter plot in the current plotting device.
The plot includes the specified variables on the axes, with labeled axes and a title.
}
\description{
This function creates a scatter plot to visualize the relationship between two specified variables
from a given data frame. The variables can be numeric and are plotted on the x-axis and y-axis.
}
\examples{
generate_plot(ourdata, "age", "systolic")

}
