% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_diff.R
\name{percent_diff}
\alias{percent_diff}
\title{Find Rows with Large Percentage Difference Between Two Columns}
\usage{
percent_diff(df, col1, col2)
}
\arguments{
\item{df}{A data frame containing the columns to compare.}

\item{col1}{A character string specifying the name of the first column.}

\item{col2}{A character string specifying the name of the second column.}
}
\value{
A data frame containing the rows where the absolute percentage difference between the
two specified columns exceeds 20\%. If no such rows exist, the function returns \code{NULL}.
}
\description{
This function calculates the absolute percentage difference between two specified columns in a
data frame and identifies the rows where the difference exceeds 20\%. If any rows have a large
difference, those rows are returned.
}
\examples{
percent_diff(ourdata, "systolic", "diastolic")

}
